+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Support go 1.18, 1.19. (@miry)
 * `toxiproxy.NewProxy` now accepts `name`, `listen addr` and `upstream addr`. (#418, @miry)
 * Replace logrus with zerolog. (#413, @miry)
-* Log HTTP requests to API server. (#413, @miry)
+* Log HTTP requests to API server. (#413, #421, @miry)
 * Add TimeoutHandler for the HTTP API server. (#420, @miry)
 
 # [2.4.0] - 2022-03-07